chore(cards): drop the now-unreferenced Ink-era primitives (#160)
CardLayout (#222) replaced everything; the old shared primitives have
no consumers left. Delete:

- src/cli/ui/primitives/CardBox.tsx (left-bar wrapper)
- src/cli/ui/primitives/BarRow.tsx (row with bar + indent)
- src/cli/ui/primitives/Pill.tsx (REASONING / OUTPUT / model badges)
- src/cli/ui/cards/CardHeader.tsx (Pill + meta header)
- src/cli/ui/cards/Card.tsx (early generic card primitive, never wired)
- src/cli/ui/cards/glyphs.ts (STRUCT.bar, STATUS.collapsed/expanded)

`CursorBlock` (the blinking-caret cell ReasoningCard uses for streaming)
moves from BarRow.tsx to its own cards/CursorBlock.tsx — it was the
only meaningful export besides BarRow itself.

`modelBadgeFor` (model-id → tier mapping) moves from Pill.tsx to
cards/model-badge.ts. Pure logic, no UI, used by ui-reasoning-tier.test.

Net diff is negative — the design language now lives entirely in
CardLayout + ProgressBar.
