+++ src/cli/ui/cards/ReasoningCard.tsx
@@ -2,11 +2,11 @@ import { Box, Text, useStdout } from "ink";
 // biome-ignore lint/style/useImportType: tsconfig jsx=react needs React in value scope for JSX compilation
 import React from "react";
 import { clipToCells, wrapToCells } from "../../../frame/width.js";
-import { CursorBlock } from "../primitives/BarRow.js";
 import { Spinner } from "../primitives/Spinner.js";
 import type { ReasoningCard as ReasoningCardData } from "../state/cards.js";
 import { FG, TONE } from "../theme/tokens.js";
 import { CardLayout } from "./CardLayout.js";
+import { CursorBlock } from "./CursorBlock.js";
 
 const STREAMING_PREVIEW_LINES = 4;
 const SETTLED_TAIL_LINES = 2;
