--- src/cli/ui/cards/glyphs.ts
@@ -1,27 +1,0 @@
-export const STATUS = {
-  ok: "✓",
-  running: "▶",
-  queued: "○",
-  failed: "✗",
-  blocked: "!",
-  collapsed: "▸",
-  expanded: "▾",
-  focused: "●",
-} as const;
-
-export const STRUCT = {
-  bar: "▎",
-  thinBar: "▏",
-  rule: "─",
-  dot: "·",
-  arrow: "›",
-} as const;
-
-export const DENSITY = {
-  full: "█",
-  high: "▓",
-  mid: "▒",
-  low: "░",
-  on: "▰",
-  off: "▱",
-} as const;
